requirements.txt
Django==3.1.1
djangorestframework==3.11.1
httpie==2.2.0
psycopg2-binary>=2.8
PyJWT==1.7.1
